Refactor
Improved class names and segregated UI components for clarity.
